% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/auxiliaries.r
\name{databinning}
\alias{databinning}
\title{Reduce a data set to representatives of bins and their frequencies}
\usage{
databinning(data, bins = 40)
}
\arguments{
\item{data}{A numeric vector of a data set.}

\item{bins}{A positive integer to represent the number of bins.}
}
\value{
A named list of \code{values} and \code{freq} whose length is \code{bins}.
}
\description{
Reduce a data set to a named list of 'values' and 'freq',
which are representatives of bins and their frequencies, respectively.
}
\examples{
rlst <- databinning(mix2gauss$n200)
}
\seealso{
\code{\link[base:rle]{base::rle()}}
}
