% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dsdp.r
\docType{package}
\name{dsdp}
\alias{dsdp}
\title{dsdp: Density Estimation using Semidefinite Programming}
\description{
The dsdp package provides three categories of important functions:
foo, bar and baz.
}
\section{dsdp functions}{

The dsdp functions ...
}

