% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/common.r
\name{estimate}
\alias{estimate}
\title{Generic Method for estimation}
\usage{
estimate(model, ...)
}
\arguments{
\item{model}{An instance of a class model to be estimated.}

\item{...}{additional arguments affecting the \code{estimate} produced.}
}
\value{
An instance of a model with estimated data.
}
\description{
This is a generic S3 method for estimation.
}
