% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/common.r
\name{func}
\alias{func}
\title{Generic Method for evaluate the estimate}
\usage{
func(model, x, ...)
}
\arguments{
\item{model}{An instance of a class model to be evaluated.}

\item{x}{A numeric vector for input.}

\item{...}{additional arguments affecting the \code{func} produced.}
}
\value{
An evaluation of \code{x} with \code{model}.
}
\description{
This is a generic S3 method for estimate.
}
