% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dataset.r
\docType{data}
\name{mix2gauss}
\alias{mix2gauss}
\title{Mixture of 2 Gauss Distributions $4.2.1}
\format{
A list of numeric vectors of Bimodal Gaussian Mixture Model.
\describe{
\item{n200}{200 elements numeric vector}
\item{n400}{400 elements numeric vector}
\item{n600}{600 elements numeric vector}
\item{n800}{800 elements numeric vector}
\item{n1000}{1000 elements numeric vector}
\item{n1200}{1200 elements numeric vector}
}
}
\source{
\code{\link{mix2gauss_gen}}
}
\usage{
mix2gauss
}
\description{
Mixture of 2 Gauss Distributions $4.2.1
}
\keyword{datasets}
