% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/auxiliaries.r
\name{mix2gauss_gen}
\alias{mix2gauss_gen}
\title{Generate mixed Gaussian random numbers}
\usage{
mix2gauss_gen(n = 100, seed = NULL)
}
\arguments{
\item{n}{The number of random numbers.}

\item{seed}{A seed for random number generator.}
}
\value{
A numeric vector of random numbers whose a density is described
in Description.
}
\description{
Generate mix gaussian random numbers whose density is
\deqn{\frac{0.3}{\sqrt{2 \pi 0.5^2}}\exp(\frac{(x+1)^2}{2 \cdot 0.5^2}) +
\frac{0.7}{\sqrt{2 \pi 0.5^2}}\exp(\frac{(x-1)^2}{2 \cdot 0.5^2}).}
}
\seealso{
mix2gauss_fun
}
