% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dataset.r
\docType{data}
\name{mix2gaussHist}
\alias{mix2gaussHist}
\title{Mixture of 2 Gauss Distributions $4.2.1, histogram version}
\format{
A list of numeric vectors of Bimodal Gaussian Mixture Model.
\describe{
\item{n200}{200 elements numeric vector}
\item{n200p}{histogram sample data with bins 25}
\item{n200f}{histogram frequency data with bins 25}
\item{n400}{400 elements numeric vector}
\item{n400p}{histogram sample data with bins 50}
\item{n400f}{histogram frequency data with bins 50}
\item{n800}{800 elements numeric vector}
\item{n800p}{histogram sample data with bins 100}
\item{n800f}{histogram frequency data with bins 100}
}
}
\source{
\code{\link{mix2gauss_gen}} \code{\link{databinning}}
}
\usage{
mix2gaussHist
}
\description{
Mixture of 2 Gauss Distributions $4.2.1, histogram version
}
\keyword{datasets}
