% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dataset.r
\docType{data}
\name{mix3gauss}
\alias{mix3gauss}
\title{Mixture of 3 Gaussian Distributions $4.2.2}
\format{
A list of numeric vectors of Unimodal Gaussian Asymmetric Mixture
Model.
\describe{
\item{n200}{200 elements numeric vector}
\item{n400}{400 elements numeric vector}
\item{n600}{600 elements numeric vector}
\item{n800}{800 elements numeric vector}
\item{n1000}{1000 elements numeric vector}
\item{n1200}{1200 elements numeric vector}
}
}
\source{
\code{\link{mix3gauss_gen}}
}
\usage{
mix3gauss
}
\description{
Mixture of 3 Gaussian Distributions $4.2.2
}
\keyword{datasets}
