% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/auxiliaries.r
\name{mix3gauss_fun}
\alias{mix3gauss_fun}
\title{A density function of mixed gaussian distribution}
\usage{
mix3gauss_fun(x)
}
\arguments{
\item{x}{A numeric vector for arguments of a density function.}
}
\value{
A numeric vector of probabilities for a given argument \code{x}.
}
\description{
A density function proportional to:
\deqn{\exp(\frac{x^2}{2}) + 5\exp(\frac{(x-1)^2}{0.2}) +
3\exp(\frac{(x-1)^2}{0.5}).}
}
\seealso{
\code{\link[=mix3gauss_gen]{mix3gauss_gen()}}
}
