% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/auxiliaries.r
\name{mix3gauss_gen}
\alias{mix3gauss_gen}
\title{Generate Mixed Gaussian Random Numbers}
\usage{
mix3gauss_gen(n = 100, seed = NULL)
}
\arguments{
\item{n}{The number of random numbers.}

\item{seed}{A seed for random number generator.}
}
\value{
A numeric vector of probabilities for a given argument \code{x}.
}
\description{
A random number generator whose density function is
proportional to:
\deqn{\exp(\frac{x^2}{2}) + 5\exp(\frac{(x-1)^2}{0.2}) +
3\exp(\frac{(x-1)^2}{0.5}).}
}
\seealso{
\code{\link[=mix3gauss_fun]{mix3gauss_fun()}}
}
