% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dataset.r
\docType{data}
\name{mixexpgamma}
\alias{mixexpgamma}
\title{Mixture of Exponential Distribution and Gamma Distribution}
\format{
A list of numeric vectors of Mixture of Exponential and Gamma
distribution Model.
\describe{
\item{n200}{200 elements numeric vector}
\item{n400}{400 elements numeric vector}
\item{n600}{600 elements numeric vector}
\item{n800}{800 elements numeric vector}
\item{n1000}{1000 elements numeric vector}
\item{n1200}{1200 elements numeric vector}
}
}
\source{
\code{\link{mixexpgamma_gen}}
}
\usage{
mixexpgamma
}
\description{
Mixture of Exponential Distribution and Gamma Distribution
}
\keyword{datasets}
