% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/auxiliaries.r
\name{mixexpgamma_fun}
\alias{mixexpgamma_fun}
\title{A density function of Mixed Exponential and Gamma Distributions}
\usage{
mixexpgamma_fun(x)
}
\arguments{
\item{x}{A numeric vector for arguments of a density function.}
}
\value{
A numeric vector of probabilities for a given argument \code{x}.
}
\description{
A density function of
\deqn{0.2(2 e^{-2x}) + 0.8 \frac{x^3}{3!}e^{-x}.}
}
\seealso{
\code{\link[=mixexpgamma_gen]{mixexpgamma_gen()}}
}
