% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/auxiliaries.r
\name{mixexpgamma_gen}
\alias{mixexpgamma_gen}
\title{Generate random numbers of Mixed Exponential and Gamma Distributions}
\usage{
mixexpgamma_gen(n = 100, seed = NULL)
}
\arguments{
\item{n}{The number of random numbers.}

\item{seed}{A seed for random number generator.}
}
\value{
A numeric vector of probabilities for a given argument \code{x}.
}
\description{
Generate random numbers whose density function:
\deqn{0.2(2 e^{-2x}) + 0.8 \frac{x^3}{3!}e^{-x}.}
}
\seealso{
\code{\link[=mixexpgamma_fun]{mixexpgamma_fun()}}
}
