% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dsdp.r
\docType{package}
\name{dsdp}
\alias{dsdp}
\title{dsdp: Density Estimation using Semidefinite Programming}
\description{
Density estimation with Semidefinite Programming.
The models of probability density functions are Gaussian or exponential
distributions with polynomial correction terms. Using a maximum likelihood
method, it computes parameters of Gaussian or exponential distributions
together with degrees of polynomials by a grid search, and coefficients
of polynomials by a variant of semidefinite programming.
It adopts Akaike Information Criterion for model selection.
See vignettes for tutorials and more information.
}
