% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dataset.r
\docType{data}
\name{mix2gauss}
\alias{mix2gauss}
\title{Datasets of Mixture of 2 Gaussian Distributions}
\format{
A list of numeric vectors of Bimodal Gaussian Mixture Model.
\describe{
\item{n200}{A numeric vector with 200 elements.}
\item{n400}{A numeric vector with 400 elements.}
\item{n600}{A numeric vector with 600 elements.}
\item{n800}{A numeric vector with 800 elements.}
\item{n1000}{A numeric vector with 1000 elements.}
\item{n1200}{A numeric vector with 1200 elements.}
}
}
\source{
\code{\link[=mix2gauss_gen]{mix2gauss_gen()}}
}
\usage{
mix2gauss
}
\description{
Dataset generated by mixture of 2 Gaussian Distributions whose density is:
\deqn{\frac{0.3}{\sqrt{2 \pi 0.5^2}}
\exp\left(\frac{(x+1)^2}{2 \cdot 0.5^2}\right) +
\frac{0.7}{\sqrt{2 \pi 0.5^2}}
\exp\left(\frac{(x-1)^2}{2 \cdot 0.5^2}\right)}.
}
\keyword{datasets}
