% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dataset.r
\docType{data}
\name{mix3gauss}
\alias{mix3gauss}
\title{Datasets of Mixture of 3 Gaussian Distributions}
\format{
A list of numeric vectors of Unimodal Gaussian Asymmetric Mixture
Model.
\describe{
\item{n200}{A numeric vector with 200 elements.}
\item{n400}{A numeric vector with 400 elements.}
\item{n600}{A numeric vector with 600 elements.}
\item{n800}{A numeric vector with 800 elements.}
\item{n1000}{A numeric vector with 1000 elements.}
\item{n1200}{A numeric vector with 1200 elements.}
}
}
\source{
\code{\link[=mix3gauss_gen]{mix3gauss_gen()}}
}
\usage{
mix3gauss
}
\description{
Datasets generated by Mixture of 3 Gaussian Distributions whose density is
proportional to:
\deqn{\exp(\frac{x^2}{2}) + 5\exp(\frac{(x-1)^2}{0.2}) +
3\exp(\frac{(x-1)^2}{0.5}).}
}
\keyword{datasets}
