% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dataset.r
\docType{data}
\name{mixexpgamma}
\alias{mixexpgamma}
\title{Dataset of Mixture of Exponential Distribution and Gamma Distribution}
\format{
A list of numeric vectors of Mixture of Exponential and Gamma
distribution Model.
\describe{
\item{n200}{A numeric vector with 200 elements.}
\item{n400}{A numeric vector with 400 elements.}
\item{n600}{A numeric vector with 600 elements.}
\item{n800}{A numeric vector with 800 elements.}
\item{n1000}{A numeric vector with 1000 elements.}
\item{n1200}{A numeric vector with 1200 elements.}
}
}
\source{
\code{\link[=mixexpgamma_gen]{mixexpgamma_gen()}}
}
\usage{
mixexpgamma
}
\description{
Dataset of mixture of exponential distribution and gamma distribution
whose density is:
\deqn{0.2(2 e^{-2x}) + 0.8 \frac{x^3}{3!}e^{-x}.}
}
\keyword{datasets}
